requirements.txt
google-cloud-aiplatform[adk,agent_engines]==1.132.0
pydantic==2.12.5
cloudpickle==3.1.2
